requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
fastapi==0.95.1
greenlet==2.0.2
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.30.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.2
pydub==0.25.1
python-multipart==0.0.6